+++ .tasks/issue/resolved/Forms_title_231218071452.yaml
@@ -6,4 +6,4 @@ description: >-
 
   Now it uses default value - form name
 comments: {}
-state: Saved
+state: Resolved
